Update Decidim::Proposals::Admin::UpdateProposal to save equity index using the concern pattern.
